requirements.txt
matplotlib
numpy
praat-parselmouth
TextGrid
